+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/connectivity/Connectivity.java
@@ -23,8 +23,8 @@ import co.elastic.apm.android.sdk.connectivity.custom.CustomConnectivity;
 import co.elastic.apm.android.sdk.connectivity.custom.CustomExporterConnectivity;
 import co.elastic.apm.android.sdk.internal.services.Service;
 import co.elastic.apm.android.sdk.internal.services.metadata.ApmMetadataService;
-import co.elastic.apm.android.sdk.internal.providers.LazyProvider;
-import co.elastic.apm.android.sdk.internal.providers.Provider;
+import co.elastic.apm.android.sdk.internal.utilities.providers.LazyProvider;
+import co.elastic.apm.android.sdk.internal.utilities.providers.Provider;
 import io.opentelemetry.sdk.logs.LogRecordProcessor;
 import io.opentelemetry.sdk.logs.export.BatchLogRecordProcessor;
 import io.opentelemetry.sdk.logs.export.LogRecordExporter;
