+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/utilities/concurrency/BackgroundWork.java
@@ -16,31 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.otel;
+package co.elastic.apm.android.sdk.internal.utilities.concurrency;
 
-import io.opentelemetry.sdk.common.CompletableResultCode;
-
-public final class Flusher {
-    private Delegator meterDelegator;
-    private Delegator loggerDelegator;
-
-    public CompletableResultCode flushMetrics() {
-        return meterDelegator.flush();
-    }
-
-    public CompletableResultCode flushLogs() {
-        return loggerDelegator.flush();
-    }
-
-    public void setMeterDelegator(Delegator meterDelegator) {
-        this.meterDelegator = meterDelegator;
-    }
-
-    public void setLoggerDelegator(Delegator loggerDelegator) {
-        this.loggerDelegator = loggerDelegator;
-    }
-
-    public interface Delegator {
-        CompletableResultCode flush();
-    }
+public interface BackgroundWork<T> {
+    T execute();
 }
