+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/utilities/otel/ElasticClock.java
@@ -16,15 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.otel;
+package co.elastic.apm.android.sdk.internal.utilities.otel;
 
 import androidx.annotation.VisibleForTesting;
 
 import java.util.concurrent.TimeUnit;
 
 import co.elastic.apm.android.common.internal.logging.Elog;
-import co.elastic.apm.android.sdk.internal.time.SystemTimeProvider;
-import co.elastic.apm.android.sdk.internal.time.ntp.TrueTimeWrapper;
+import co.elastic.apm.android.sdk.internal.utilities.time.SystemTimeProvider;
+import co.elastic.apm.android.sdk.internal.utilities.time.ntp.TrueTimeWrapper;
 import io.opentelemetry.sdk.common.Clock;
 
 public class ElasticClock implements Clock {
