+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/utilities/time/ntp/TrueTimeBootBroadcastReceiver.java
@@ -16,9 +16,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.providers;
+package co.elastic.apm.android.sdk.internal.utilities.time.ntp;
 
-public interface Provider<T> {
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
 
-    T get();
+import co.elastic.apm.android.common.internal.logging.Elog;
+
+public class TrueTimeBootBroadcastReceiver extends BroadcastReceiver {
+
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        Elog.getLogger().debug("Clearing TrueTime disk cache as we've detected a boot");
+        if (Intent.ACTION_BOOT_COMPLETED.equals(intent.getAction())) {
+            TrueTimeWrapper trueTimeWrapper = new TrueTimeWrapper(context);
+            trueTimeWrapper.withSharedPreferencesCache();
+            trueTimeWrapper.clearCachedInfo();
+        }
+    }
 }
